web4-policy 2a: extract the generic Law engine; hub re-consumes as Law<HubPolicy>
RFC #419 step-2, phase 2a. Lifts the domain-agnostic R6 policy engine out of
hub-lib/src/law.rs into a new `web4-policy` crate (web4-core -> web4-policy ->
{hub, hestia, hardbound}), per Legion's Q5 TYPED verdict (Law<E: PolicyExtension>,
verified to fit hestia ConstellationPolicy + hardbound RBACExtension).

web4-policy (generic): Law<E>, Norm/Decision/Operator/Procedure/Condition/
R6Request/DecisionOutcome, evaluate/evaluate_outcome (priority conflict
resolution, default-Allow), structural validate (rules 1-5 + escalation shape),
the hydrate mechanism (delegates the *what* to E), serde + sha256. Identity/
role-trust is a separate surface (Q2), not in E — deferred to 2b behind RFC #403.

hub-lib (society specialization): HubPolicy{admission,delegation,atp_issuance}:
PolicyExtension — validate() carries rules 6-10 + KNOWN_ROLES; hydrate_defaults()
the #417 admission defaults. `type Law = web4_policy::Law<HubPolicy>`; generic
types re-exported so existing imports are unchanged. Admission accessors move to
a HubLawExt trait (Law is now a foreign alias).

Wire-neutral: HubPolicy is `#[serde(flatten)]`-ed so the law's on-disk/over-wire
shape is unchanged — proven by a canary test parsing the EXACT live HUB law
(v1.0.2) and asserting fixed-point serialization (no hash drift), correct
evaluation, and admission accessors. Behavior-neutral: 32 daemon + 157 lib + 5
web4-policy tests green (pre-extraction counts preserved + new engine/canary tests).

Co-owned with Legion (hestia/hardbound). 2b (composite-identity/role-trust)
trails this behind RFC #403 P2 <-> mcp-protocol §7.5 reconciliation.
