+++ hub/Cargo.toml
@@ -14,6 +14,7 @@ description = "Web4 Community Hub — minimum-viable Web4 society for a communit
 # Web4 primitives — DO NOT reimplement. Use these directly.
 web4-core = { path = "../web4-core" }
 web4-trust-core = { path = "../web4-trust-core" }
+web4-policy = { path = "../web4-policy" }
 
 # Common third-party
 anyhow = "1"
